All tests are passing

+++ rbush/data/__init__.py
@@ -7,8 +7,8 @@ def generate_data(n,size=50):
     for _ in range(n):
         xmin = int(random() * (100-size))
         ymin = int(random() * (100-size))
-        xmax = xmin + size * random()
-        ymax = ymin + size * random()
+        xmax = int(xmin + size * random())
+        ymax = int(ymin + size * random())
         item = {'xmin': xmin, 'ymin': ymin, 'xmax': xmax, 'ymax': ymax}
         data.append(item)
     return data
